Allow the user to add and remove tags on a package
And implement the saving of information about a package in the
installed_packages.list file. It's a bit hackish, but it works,
thanks a "positive bug" of databasewriter.cpp :

The content between [ and ] in installed_packages.list is
ignored, and only Name and Version are used to identify the
package. An uniq ID is used between [ and ], and the Name
and Version attributes are set correctly.

On Setup Update, [blabla] is ignored, Name and Version are used
to find the package, and Flags is read. It's done, and it's very
simple !