Support for orphaned packages
It's a really nice feature requested by many user of well-known package
managers.

A new tag, wanted, has been added to Setup. If it's set, a package
cannot be removed if all the packages depending on it are deleted.

A counter is used. It retains the number of installed packages using
another package. If this counter goes to 0, and the wanted tag of the
package is not set, the package is considered as being orphaned and can
be automatically removed with "setup autoremove".

Thanks to that, the user's system is always clean.