Use a way clever algorithm to export the list of the files, 37 times faster !
Instead of querying every directory and building a list of files in it, we
make a big query, asking MySQL to return lines like :

[ /etc/initng/daemon/slim | <flags> | slim ]

They are sorted. We can now use the simple and fast code from cache.cpp, and
change it not to display a tree, but to display the files.xz contents.

setup update took 3 minutes 26 seconds, it takes now 5,45 seconds !