Installation, removing and updating of the packages
It's the so wanted commit ! Setup can now install, remove and update
packages without any problems, and way faster than before for the
packages containing many files (we don't have to parse a giant plain
text file containing the list of the files).

The flags of the files are fully supported.