Handle the nodes being explored in Solver
Fixes the bug which does that bash can be installed without libreadline6, when we use "lpm add bash=4.1~2".

Solver::verifyNode was wrong when he checked for Node::Explored, a flag that is also set if we restart the solving after a choice.

Some changes were also added to Solver::upList to fixe some glitches in the choices.