It's now PackageSource that embeds the icon.
It's way cleaner, less hackish, and we will be able to handle SVG icons. The packagers can specify an exact file name to be embedded (filename that is kept in the metadata once the package is built), they're not obliged to put their icons under control/icon.* , and can even embed it themselves as a CDATA section (SVG text).

The icon attribute can have an type="svg" attribute in the future, to allow SVG icons (that aren't detected using QImage::fromData, and need to use QSvgRenderer).

This commit makes the one taling about the fallback icon in Pkgui working.